Move my_print_defaults to mariadb-server-core (Closes: #898367)
The command is needed to install/upgrade databases. It should therefore
reside in the server-core package, as that package is supposed to contain
all necessary tools to start and run the database (although manually,
there are no automation or service scripts included in server-core by
design).

Breaks+Replaces in the control file already exists since previous file
moves so they do not need to be re-created.