Rename 'mariadbcheck' to 'mariadb-check' as upstream is doing in 10.4
It is better to stay aligned with upstream in these names so we don't
divert in downstream Debian. Checking for external dependencies there
does not seem to be anything that is fixed on this name, and we can
simply proceed with the rename now.

For example https://codesearch.debian.net/search?q=mariadbcheck does not
show any external use of the name 'mariadbcheck' (yet!).