Make mariadb-client-10.4 Recommends libdbd-mariadb-perl as primary option
The package libdbd-mariadb-perl has been available in Debian September
2018, so it is already included since Debian Buster and Ubuntu Disco.

Use it as the primary Recommends for MariaDB client binaries, but keep
the MySQL variant as a fall-back in case this package gets backported to
older releases that did not yet have it.

Drop the version for libdbd-mysql-perl since version 2.1020 that
superseded version 1.2219 was released in 2002 and there is no need to
keep carrying over that information.

https://tracker.debian.org/pkg/libdbd-mysql-perl
https://tracker.debian.org/pkg/libdbd-mariadb-perl