Simplify autopkgtest 'smoke' to be easier to debug
Simplify Debian autopkgtest RocksDB part and make it more verbose so that
future regressions like this are easier to debug.

Also remove QUIET from the RocksDB FIND_PACKAGE call so that it is easier
to read in build logs what libraries were detected. Also add missing
underscores to error messages.

# Conflicts:
#	debian/tests/smoke