Make SysV init script explicit on its dependencies (Related: #1035949)
In MDEV-28640 the init script failed to stop/start the MariaDB server
due to missing mysqladmin on the system. This was however very hard to
spot from the console output.

Add an explicit check for the binary the script depends on, and fail
verbosely if the dependency is missing.

This is identical as done in upstream
https://github.com/MariaDB/server/pull/2299/commits/a7fb9f80af6f6aa2e5ca2f36a79d325358b40e00
and later in Debian in
https://salsa.debian.org/mariadb-team/mariadb-server/-/commit/931c1c6c0734bd85fea025294fdf4edeeeb7970c
with the addition of using '-x' instead of '-f' in check.

Cherry-picked from 42ac08f6.